replace index.php with index.html
The index doesn’t have to be a PHP file since there is no need to
generate the list of chapters anew every time the ToC is opened.

Instead, the list of chapters is generated once, ahead of time, and
saved to the index.html.

As an added bonus, this allows the ToC to be generated and used
as a local file.