style: Ass <span class="snon-speech"> and make <span class="thought"> red in preparation for its removal
I went a bit overboard with literal thought quotes and it didn't even fit
all my use cases so I decided to replace it with an alternative that better
fits what I want to use it for.

TODO: Change the solid triangle on <span class="speech"> for the open ones
currently used for <span class="non-speech"> and find a new symbol to use
for non-speech.