'ma' Escape: revision of the automated announcements on the train
I wrote German versions of the announcements (look at the code if you wanna
see them) and had a back and forth with ChatGPT to improve the English ones.

Note: I am not using ChatGPT to write or even help me write. But it works
great to figure out language intuition and vocabulary because English isn't
my native language.